requirements.txt
timm==0.4.12
albumentations==1.0.0
neptune-client==0.9.16
transformers==4.6.1
geffnet==1.0.2
efficientnet-pytorch==0.7.1
